backend/requirements.txt
aiodns==3.2.0
aiofile==3.8.8
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
boto3==1.34.151
botocore==1.34.151
cachelib==0.13.0
caio==0.9.17
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.0.5
psycog==0.0.1
psycopg2==2.9.9
pycares==4.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
s3transfer==0.10.2
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.31
stream-chat==4.18.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
